Create custom type for K8SResource when run in serverless.
